+++ driver-async/src/main/com/mongodb/async/client/MongoClients.java
@@ -17,7 +17,7 @@
 package com.mongodb.async.client;
 
 import com.mongodb.ConnectionString;
-import com.mongodb.client.gridfs.model.GridFSFileCodecProvider;
+import com.mongodb.client.gridfs.codecs.GridFSFileCodecProvider;
 import com.mongodb.client.model.geojson.codecs.GeoJsonCodecProvider;
 import com.mongodb.connection.Cluster;
 import com.mongodb.connection.ClusterSettings;
