+++ driver/src/main/com/mongodb/MongoClient.java
@@ -19,7 +19,7 @@ package com.mongodb;
 import com.mongodb.client.ListDatabasesIterable;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.MongoIterable;
-import com.mongodb.client.gridfs.model.GridFSFileCodecProvider;
+import com.mongodb.client.gridfs.codecs.GridFSFileCodecProvider;
 import com.mongodb.client.model.geojson.codecs.GeoJsonCodecProvider;
 import org.bson.BsonDocument;
 import org.bson.Document;
